StorageReader exposes new read interfaces
as well as being a CursorFactory, which allows a generic
implementation of ReadOperations/WriteOperations in the kernel
which isn't tied to a specific storage implementation.
