+++ community/kernel/src/main/java/org/neo4j/kernel/impl/newapi/WriteOperations.java
@@ -29,18 +29,16 @@ import java.util.Map;
 import org.neo4j.helpers.collection.CastingIterator;
 import org.neo4j.helpers.collection.Iterators;
 import org.neo4j.internal.kernel.api.CapableIndexReference;
-import org.neo4j.internal.kernel.api.CursorFactory;
-import org.neo4j.internal.kernel.api.ExplicitIndexRead;
 import org.neo4j.internal.kernel.api.ExplicitIndexWrite;
 import org.neo4j.internal.kernel.api.IndexQuery;
 import org.neo4j.internal.kernel.api.IndexReference;
 import org.neo4j.internal.kernel.api.Locks;
+import org.neo4j.internal.kernel.api.NodeCursor;
 import org.neo4j.internal.kernel.api.NodeLabelIndexCursor;
-import org.neo4j.internal.kernel.api.Procedures;
-import org.neo4j.internal.kernel.api.Read;
-import org.neo4j.internal.kernel.api.SchemaRead;
+import org.neo4j.internal.kernel.api.NodeValueIndexCursor;
+import org.neo4j.internal.kernel.api.PropertyCursor;
+import org.neo4j.internal.kernel.api.RelationshipScanCursor;
 import org.neo4j.internal.kernel.api.SchemaWrite;
-import org.neo4j.internal.kernel.api.Token;
 import org.neo4j.internal.kernel.api.Write;
 import org.neo4j.internal.kernel.api.exceptions.EntityNotFoundException;
 import org.neo4j.internal.kernel.api.exceptions.KernelException;
@@ -56,7 +54,6 @@ import org.neo4j.internal.kernel.api.schema.constraints.ConstraintDescriptor;
 import org.neo4j.kernel.api.SilentTokenNameLookup;
 import org.neo4j.kernel.api.StatementConstants;
 import org.neo4j.kernel.api.exceptions.index.IndexEntryConflictException;
-import org.neo4j.kernel.api.exceptions.index.IndexNotApplicableKernelException;
 import org.neo4j.kernel.api.exceptions.index.IndexNotFoundKernelException;
 import org.neo4j.kernel.api.exceptions.schema.AlreadyConstrainedException;
 import org.neo4j.kernel.api.exceptions.schema.AlreadyIndexedException;
@@ -84,9 +81,11 @@ import org.neo4j.kernel.impl.api.KernelTransactionImplementation;
 import org.neo4j.kernel.impl.api.state.ConstraintIndexCreator;
 import org.neo4j.kernel.impl.api.store.DefaultIndexReference;
 import org.neo4j.kernel.impl.constraints.ConstraintSemantics;
+import org.neo4j.kernel.impl.index.ExplicitIndexStore;
 import org.neo4j.kernel.impl.index.IndexEntityType;
 import org.neo4j.kernel.impl.locking.ResourceTypes;
 import org.neo4j.storageengine.api.EntityType;
+import org.neo4j.storageengine.api.StorageEngine;
 import org.neo4j.storageengine.api.StorageReader;
 import org.neo4j.storageengine.api.lock.ResourceType;
 import org.neo4j.values.storable.Value;
@@ -100,65 +99,66 @@ import static org.neo4j.internal.kernel.api.schema.SchemaDescriptorPredicates.ha
 import static org.neo4j.kernel.api.StatementConstants.NO_SUCH_NODE;
 import static org.neo4j.kernel.api.StatementConstants.NO_SUCH_PROPERTY_KEY;
 import static org.neo4j.kernel.api.schema.index.SchemaIndexDescriptor.Type.UNIQUE;
+import static org.neo4j.kernel.impl.api.store.DefaultIndexReference.toDescriptor;
 import static org.neo4j.kernel.impl.locking.ResourceTypes.INDEX_ENTRY;
 import static org.neo4j.kernel.impl.locking.ResourceTypes.indexEntryResourceId;
 import static org.neo4j.kernel.impl.newapi.IndexTxStateUpdater.LabelChangeType.ADDED_LABEL;
 import static org.neo4j.kernel.impl.newapi.IndexTxStateUpdater.LabelChangeType.REMOVED_LABEL;
 import static org.neo4j.values.storable.Values.NO_VALUE;
 
-
 /**
  * Collects all Kernel API operations and guards them from being used outside of transaction.
+ * This class is designed to be generic, even for different {@link StorageEngine}.
+ * It uses {@link ReadOperations} and {@link StorageReader} to encode logical changes into {@link TransactionState}.
+ * Changes also generally involve first {@link Locks locking} the modified resources.
  */
-public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
+public class WriteOperations implements Write, ExplicitIndexWrite, SchemaWrite
 {
     private final KernelTransactionImplementation ktx;
-    private final AllStoreHolder allStoreHolder;
     private final KernelToken token;
-    private final StorageReader statement;
+    private final StorageReader read;
     private final AutoIndexing autoIndexing;
-    private DefaultNodeCursor nodeCursor;
     private final IndexTxStateUpdater updater;
-    private DefaultPropertyCursor propertyCursor;
-    private DefaultRelationshipScanCursor relationshipCursor;
-    private final DefaultCursors cursors;
     private final ConstraintIndexCreator constraintIndexCreator;
     private final ConstraintSemantics constraintSemantics;
+    private final ExplicitIndexStore explicitIndexStore;
 
-    public Operations(
-            AllStoreHolder allStoreHolder,
+    private NodeCursor nodeCursor;
+    private PropertyCursor propertyCursor;
+    private RelationshipScanCursor relationshipCursor;
+
+    public WriteOperations(
             IndexTxStateUpdater updater,
-            StorageReader statement,
+            StorageReader read,
             KernelTransactionImplementation ktx,
             KernelToken token,
-            DefaultCursors cursors,
             AutoIndexing autoIndexing,
             ConstraintIndexCreator constraintIndexCreator,
-            ConstraintSemantics constraintSemantics )
+            ConstraintSemantics constraintSemantics,
+            ExplicitIndexStore explicitIndexStore )
     {
         this.token = token;
         this.autoIndexing = autoIndexing;
-        this.allStoreHolder = allStoreHolder;
         this.ktx = ktx;
-        this.statement = statement;
+        this.read = read;
         this.updater = updater;
-        this.cursors = cursors;
         this.constraintIndexCreator = constraintIndexCreator;
         this.constraintSemantics = constraintSemantics;
+        this.explicitIndexStore = explicitIndexStore;
     }
 
     public void initialize()
     {
-        this.nodeCursor = cursors.allocateNodeCursor();
-        this.propertyCursor = cursors.allocatePropertyCursor();
-        this.relationshipCursor = cursors.allocateRelationshipScanCursor();
+        this.nodeCursor = read.allocateNodeCursor();
+        this.propertyCursor = read.allocatePropertyCursor();
+        this.relationshipCursor = read.allocateRelationshipScanCursor();
     }
 
     @Override
     public long nodeCreate()
     {
         ktx.assertOpen();
-        long nodeId = statement.reserveNode();
+        long nodeId = read.reserveNode();
         ktx.txState().nodeDoCreate( nodeId );
         return nodeId;
     }
@@ -204,7 +204,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         assertNodeExists( sourceNode );
         assertNodeExists( targetNode );
 
-        long id = statement.reserveRelationship();
+        long id = read.reserveRelationship();
         ktx.txState().relationshipDoCreate( id, relationshipType, sourceNode, targetNode );
         return id;
     }
@@ -235,7 +235,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         //Check so that we are not breaking uniqueness constraints
         //We do this by checking if there is an existing node in the index that
         //with the same label and property combination.
-        Iterator<ConstraintDescriptor> constraints = allStoreHolder.constraintsGetForLabel( nodeLabel );
+        Iterator<ConstraintDescriptor> constraints = read.constraintsGetForLabel( nodeLabel );
         while ( constraints.hasNext() )
         {
             ConstraintDescriptor constraint = constraints.next();
@@ -280,7 +280,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         {
             ktx.statementLocks().optimistic().acquireExclusive( ktx.lockTracer(), ResourceTypes.NODE, node );
         }
-        if ( allStoreHolder.nodeExistsInStore( node ) )
+        if ( read.nodeExists( node ) )
         {
             autoIndexing.nodes().entityRemoved( this, node );
             ktx.txState().nodeDoDelete( node );
@@ -293,7 +293,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
 
     private boolean relationshipDelete( long relationship, boolean lock ) throws AutoIndexingKernelException
     {
-        allStoreHolder.singleRelationship( relationship, relationshipCursor ); // tx-state aware
+        read.singleRelationship( relationship, relationshipCursor ); // tx-state aware
 
         if ( relationshipCursor.next() )
         {
@@ -303,7 +303,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
                         relationshipCursor.targetNodeReference() );
                 acquireExclusiveRelationshipLock( relationship );
             }
-            if ( !allStoreHolder.relationshipExists( relationship ) )
+            if ( !read.relationshipExists( relationship ) )
             {
                 return false;
             }
@@ -319,7 +319,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             }
             else
             {
-                txState.relationshipDoDelete( relationship, relationshipCursor.getType(),
+                txState.relationshipDoDelete( relationship, relationshipCursor.type(),
                         relationshipCursor.sourceNodeReference(), relationshipCursor.targetNodeReference() );
             }
             return true;
@@ -331,7 +331,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
 
     private void singleNode( long node ) throws EntityNotFoundException
     {
-        allStoreHolder.singleNode( node, nodeCursor );
+        read.singleNode( node, nodeCursor );
         if ( !nodeCursor.next() )
         {
             throw new EntityNotFoundException( EntityType.NODE, node );
@@ -340,7 +340,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
 
     private void singleRelationship( long relationship ) throws EntityNotFoundException
     {
-        allStoreHolder.singleRelationship( relationship, relationshipCursor );
+        read.singleRelationship( relationship, relationshipCursor );
         if ( !relationshipCursor.next() )
         {
             throw new EntityNotFoundException( EntityType.RELATIONSHIP, relationship );
@@ -395,14 +395,15 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
      * Check so that there is not an existing node with the exact match of label and property
      */
     private void validateNoExistingNodeWithExactValues( IndexBackedConstraintDescriptor constraint,
-            IndexQuery.ExactPredicate[] propertyValues, long modifiedNode
-    ) throws UniquePropertyValueValidationException, UnableToValidateConstraintException
+            IndexQuery.ExactPredicate[] propertyValues, long modifiedNode )
+            throws UniquePropertyValueValidationException, UnableToValidateConstraintException
     {
-        try ( DefaultNodeValueIndexCursor valueCursor = cursors.allocateNodeValueIndexCursor() )
+        try ( NodeValueIndexCursor valueCursor = read.allocateNodeValueIndexCursor() )
         {
             SchemaIndexDescriptor schemaIndexDescriptor = constraint.ownedIndexDescriptor();
             assertIndexOnline( schemaIndexDescriptor );
-            int labelId = schemaIndexDescriptor.schema().keyId();
+            SchemaDescriptor schema = schemaIndexDescriptor.schema();
+            int labelId = schema.keyId();
 
             //Take a big fat lock, and check for existing node in index
             ktx.statementLocks().optimistic().acquireExclusive(
@@ -410,16 +411,16 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
                     indexEntryResourceId( labelId, propertyValues )
             );
 
-            allStoreHolder.nodeIndexSeekWithFreshIndexReader(
-                    allStoreHolder.indexGetCapability( schemaIndexDescriptor ), valueCursor, propertyValues );
-            if ( valueCursor.next() && valueCursor.nodeReference() != modifiedNode )
+            CapableIndexReference index = read.index( labelId, schema.getPropertyIds() );
+            long existingNode = read.lockingNodeUniqueIndexSeek( index, propertyValues );
+            if ( existingNode != modifiedNode )
             {
                 throw new UniquePropertyValueValidationException( constraint, VALIDATION,
-                        new IndexEntryConflictException( valueCursor.nodeReference(), NO_SUCH_NODE,
+                        new IndexEntryConflictException( existingNode, NO_SUCH_NODE,
                                 IndexQuery.asValueTuple( propertyValues ) ) );
             }
         }
-        catch ( IndexNotFoundKernelException | IndexBrokenKernelException | IndexNotApplicableKernelException e )
+        catch ( KernelException e )
         {
             throw new UnableToValidateConstraintException( constraint, e );
         }
@@ -428,12 +429,12 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
     private void assertIndexOnline( SchemaIndexDescriptor descriptor )
             throws IndexNotFoundKernelException, IndexBrokenKernelException
     {
-        switch ( allStoreHolder.indexGetState( descriptor ) )
+        switch ( read.indexGetState( descriptor ) )
         {
         case ONLINE:
             return;
         default:
-            throw new IndexBrokenKernelException( allStoreHolder.indexGetFailure( descriptor ) );
+            throw new IndexBrokenKernelException( read.indexGetFailure( descriptor.schema() ) );
         }
     }
 
@@ -468,7 +469,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         ktx.statementLocks().optimistic().acquireShared( ktx.lockTracer(), ResourceTypes.LABEL,
                 nodeCursor.labels().all() );
         Iterator<ConstraintDescriptor> constraints = Iterators.filter( hasProperty( propertyKey ),
-                allStoreHolder.constraintsGetAll() );
+                read.constraintsGetAll() );
         Iterator<IndexBackedConstraintDescriptor> uniquenessConstraints =
                 new CastingIterator<>( constraints, IndexBackedConstraintDescriptor.class );
 
@@ -609,14 +610,14 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        allStoreHolder.explicitIndexTxState().nodeChanges( indexName ).addNode( node, key, value );
+        ktx.explicitIndexTxState().nodeChanges( indexName ).addNode( node, key, value );
     }
 
     @Override
     public void nodeRemoveFromExplicitIndex( String indexName, long node ) throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        allStoreHolder.explicitIndexTxState().nodeChanges( indexName ).remove( node );
+        ktx.explicitIndexTxState().nodeChanges( indexName ).remove( node );
     }
 
     @Override
@@ -624,7 +625,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        allStoreHolder.explicitIndexTxState().nodeChanges( indexName ).remove( node, key, value );
+        ktx.explicitIndexTxState().nodeChanges( indexName ).remove( node, key, value );
     }
 
     @Override
@@ -632,28 +633,28 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        allStoreHolder.explicitIndexTxState().nodeChanges( indexName ).remove( node, key );
+        ktx.explicitIndexTxState().nodeChanges( indexName ).remove( node, key );
     }
 
     @Override
     public void nodeExplicitIndexCreate( String indexName, Map<String,String> customConfig )
     {
         ktx.assertOpen();
-        allStoreHolder.explicitIndexTxState().createIndex( IndexEntityType.Node, indexName, customConfig );
+        ktx.explicitIndexTxState().createIndex( IndexEntityType.Node, indexName, customConfig );
     }
 
     @Override
     public void nodeExplicitIndexCreateLazily( String indexName, Map<String,String> customConfig )
     {
         ktx.assertOpen();
-        allStoreHolder.getOrCreateNodeIndexConfig( indexName, customConfig );
+        explicitIndexStore.getOrCreateNodeIndexConfig( indexName, customConfig );
     }
 
     @Override
     public void nodeExplicitIndexDrop( String indexName ) throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        ExplicitIndexTransactionState txState = allStoreHolder.explicitIndexTxState();
+        ExplicitIndexTransactionState txState = ktx.explicitIndexTxState();
         txState.nodeChanges( indexName ).drop();
         txState.deleteIndex( IndexEntityType.Node, indexName );
     }
@@ -663,7 +664,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
        ktx.assertOpen();
-       return allStoreHolder.explicitIndexStore().setNodeIndexConfiguration( indexName, key, value );
+       return explicitIndexStore.setNodeIndexConfiguration( indexName, key, value );
     }
 
     @Override
@@ -671,7 +672,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        return allStoreHolder.explicitIndexStore().removeNodeIndexConfiguration( indexName, key );
+        return explicitIndexStore.removeNodeIndexConfiguration( indexName, key );
     }
 
     @Override
@@ -679,10 +680,10 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException, EntityNotFoundException
     {
         ktx.assertOpen();
-        allStoreHolder.singleRelationship( relationship, relationshipCursor );
+        read.singleRelationship( relationship, relationshipCursor );
         if ( relationshipCursor.next() )
         {
-            allStoreHolder.explicitIndexTxState().relationshipChanges( indexName ).addRelationship( relationship, key, value,
+            ktx.explicitIndexTxState().relationshipChanges( indexName ).addRelationship( relationship, key, value,
                     relationshipCursor.sourceNodeReference(), relationshipCursor.targetNodeReference() );
         }
         else
@@ -696,7 +697,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        allStoreHolder.explicitIndexTxState().relationshipChanges( indexName ).remove( relationship, key, value );
+        ktx.explicitIndexTxState().relationshipChanges( indexName ).remove( relationship, key, value );
     }
 
     @Override
@@ -704,7 +705,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        allStoreHolder.explicitIndexTxState().relationshipChanges( indexName ).remove( relationship, key );
+        ktx.explicitIndexTxState().relationshipChanges( indexName ).remove( relationship, key );
 
     }
 
@@ -713,28 +714,28 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        allStoreHolder.explicitIndexTxState().relationshipChanges( indexName ).remove( relationship );
+        ktx.explicitIndexTxState().relationshipChanges( indexName ).remove( relationship );
     }
 
     @Override
     public void relationshipExplicitIndexCreate( String indexName, Map<String,String> customConfig )
     {
         ktx.assertOpen();
-        allStoreHolder.explicitIndexTxState().createIndex( IndexEntityType.Relationship, indexName, customConfig );
+        ktx.explicitIndexTxState().createIndex( IndexEntityType.Relationship, indexName, customConfig );
     }
 
     @Override
     public void relationshipExplicitIndexCreateLazily( String indexName, Map<String,String> customConfig )
     {
         ktx.assertOpen();
-        allStoreHolder.getOrCreateRelationshipIndexConfig( indexName, customConfig );
+        explicitIndexStore.getOrCreateRelationshipIndexConfig( indexName, customConfig );
     }
 
     @Override
     public void relationshipExplicitIndexDrop( String indexName ) throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        ExplicitIndexTransactionState txState = allStoreHolder.explicitIndexTxState();
+        ExplicitIndexTransactionState txState = ktx.explicitIndexTxState();
         txState.relationshipChanges( indexName ).drop();
         txState.deleteIndex( IndexEntityType.Relationship, indexName );
     }
@@ -775,7 +776,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
 
     private Value readGraphProperty( int propertyKey )
     {
-        allStoreHolder.graphProperties( propertyCursor );
+        read.graphProperties( propertyCursor );
 
         //Find out if the property had a value
         Value existingValue = NO_VALUE;
@@ -788,16 +789,6 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             }
         }
         return existingValue;
-    }
-
-    public CursorFactory cursors()
-    {
-        return cursors;
-    }
-
-    public Procedures procedures()
-    {
-        return allStoreHolder;
     }
 
     public void release()
@@ -817,44 +808,6 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             relationshipCursor.close();
             relationshipCursor = null;
         }
-
-        cursors.assertClosed();
-        cursors.release();
-    }
-
-    public Token token()
-    {
-        return token;
-    }
-
-    public ExplicitIndexRead indexRead()
-    {
-        return allStoreHolder;
-    }
-
-    public SchemaRead schemaRead()
-    {
-        return allStoreHolder;
-    }
-
-    public Read dataRead()
-    {
-        return allStoreHolder;
-    }
-
-    public DefaultNodeCursor nodeCursor()
-    {
-        return nodeCursor;
-    }
-
-    public DefaultRelationshipScanCursor relationshipCursor()
-    {
-        return relationshipCursor;
-    }
-
-    public DefaultPropertyCursor propertyCursor()
-    {
-        return propertyCursor;
     }
 
     @Override
@@ -880,7 +833,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         org.neo4j.kernel.api.schema.LabelSchemaDescriptor descriptor = labelDescriptor( index );
         try
         {
-            SchemaIndexDescriptor existingIndex = allStoreHolder.indexGetForSchema( descriptor );
+            SchemaIndexDescriptor existingIndex = read.indexGetForSchema( descriptor );
 
             if ( existingIndex == null )
             {
@@ -889,7 +842,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
 
             if ( existingIndex.type() == UNIQUE )
             {
-                if ( allStoreHolder.indexGetOwningUniquenessConstraintId( existingIndex ) != null )
+                if ( read.indexGetOwningUniquenessConstraintId( existingIndex ) != null )
                 {
                     throw new IndexBelongsToConstraintException( descriptor );
                 }
@@ -899,7 +852,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         {
             throw new DropIndexFailureException( descriptor, e );
         }
-        ktx.txState().indexDoDrop( allStoreHolder.indexDescriptor( index ) );
+        ktx.txState().indexDoDrop( toDescriptor( index ) );
     }
 
     @Override
@@ -937,9 +890,9 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         assertIndexDoesNotExist( SchemaKernelException.OperationContext.CONSTRAINT_CREATION, descriptor );
 
         //enforce constraints
-        try ( NodeLabelIndexCursor nodes = cursors.allocateNodeLabelIndexCursor() )
+        try ( NodeLabelIndexCursor nodes = read.allocateNodeLabelIndexCursor() )
         {
-            allStoreHolder.nodeLabelScan( descriptor.getLabelId(), nodes );
+            read.nodeLabelScan( descriptor.getLabelId(), nodes );
             constraintSemantics.validateNodeKeyConstraint( nodes, nodeCursor, propertyCursor, descriptor );
         }
 
@@ -962,9 +915,9 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         assertConstraintDoesNotExist( constraint );
 
         //enforce constraints
-        try ( NodeLabelIndexCursor nodes = cursors.allocateNodeLabelIndexCursor() )
+        try ( NodeLabelIndexCursor nodes = read.allocateNodeLabelIndexCursor() )
         {
-            allStoreHolder.nodeLabelScan( descriptor.getLabelId(), nodes );
+            read.nodeLabelScan( descriptor.getLabelId(), nodes );
             constraintSemantics
                     .validateNodePropertyExistenceConstraint( nodes, nodeCursor, propertyCursor, descriptor );
         }
@@ -988,7 +941,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         assertConstraintDoesNotExist( constraint );
 
         //enforce constraints
-        allStoreHolder.relationshipTypeScan( descriptor.getRelTypeId(), relationshipCursor );
+        read.relationshipTypeScan( descriptor.getRelTypeId(), relationshipCursor );
         constraintSemantics
                 .validateRelationshipPropertyExistenceConstraint( relationshipCursor, propertyCursor, descriptor );
 
@@ -1003,7 +956,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        return allStoreHolder.explicitIndexStore().setRelationshipIndexConfiguration( indexName, key, value );
+        return explicitIndexStore.setRelationshipIndexConfiguration( indexName, key, value );
     }
 
     @Override
@@ -1011,7 +964,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             throws ExplicitIndexNotFoundKernelException
     {
         ktx.assertOpen();
-        return allStoreHolder.explicitIndexStore().removeRelationshipIndexConfiguration( indexName, key );
+        return explicitIndexStore.removeRelationshipIndexConfiguration( indexName, key );
     }
 
     @Override
@@ -1039,7 +992,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
     private void assertIndexDoesNotExist( SchemaKernelException.OperationContext context,
             SchemaDescriptor descriptor ) throws AlreadyIndexedException, AlreadyConstrainedException
     {
-        SchemaIndexDescriptor existingIndex = allStoreHolder.indexGetForSchema( descriptor );
+        SchemaIndexDescriptor existingIndex = read.indexGetForSchema( descriptor );
         if ( existingIndex != null )
         {
             // OK so we found a matching constraint index. We check whether or not it has an owner
@@ -1122,7 +1075,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
 
     private void assertNodeExists( long sourceNode ) throws EntityNotFoundException
     {
-        if ( !allStoreHolder.nodeExists( sourceNode ) )
+        if ( !read.nodeExists( sourceNode ) )
         {
             throw new EntityNotFoundException( EntityType.NODE, sourceNode );
         }
@@ -1130,13 +1083,13 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
 
     private boolean constraintIndexHasOwner( SchemaIndexDescriptor descriptor )
     {
-        return allStoreHolder.indexGetOwningUniquenessConstraintId( descriptor ) != null;
+        return read.indexGetOwningUniquenessConstraintId( descriptor ) != null;
     }
 
     private void assertConstraintDoesNotExist( ConstraintDescriptor constraint )
             throws AlreadyConstrainedException
     {
-        if ( allStoreHolder.constraintExists( constraint ) )
+        if ( read.constraintExists( constraint ) )
         {
             throw new AlreadyConstrainedException( constraint,
                     SchemaKernelException.OperationContext.CONSTRAINT_CREATION,
@@ -1144,15 +1097,10 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         }
     }
 
-    public Locks locks()
-    {
-        return allStoreHolder;
-    }
-
     private void assertConstraintExists( ConstraintDescriptor constraint )
             throws NoSuchConstraintException
     {
-        if ( !allStoreHolder.constraintExists( constraint ) )
+        if ( !read.constraintExists( constraint ) )
         {
             throw new NoSuchConstraintException( constraint );
         }
@@ -1190,7 +1138,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
             }
             else // *CREATE*
             { // create from scratch
-                Iterator<ConstraintDescriptor> it = allStoreHolder.constraintsGetForSchema( descriptor );
+                Iterator<ConstraintDescriptor> it = read.constraintsGetForSchema( descriptor );
                 while ( it.hasNext() )
                 {
                     if ( it.next().equals( constraint ) )
@@ -1199,7 +1147,7 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
                     }
                 }
                 long indexId = constraintIndexCreator.createUniquenessConstraintIndex( ktx, descriptor );
-                if ( !allStoreHolder.constraintExists( constraint ) )
+                if ( !read.constraintExists( constraint ) )
                 {
                     // This looks weird, but since we release the label lock while awaiting population of the index
                     // backing this constraint there can be someone else getting ahead of us, creating this exact
@@ -1223,5 +1171,20 @@ public class Operations implements Write, ExplicitIndexWrite, SchemaWrite
         {
             throw new NoSuchIndexException( SchemaDescriptorFactory.forLabel( index.label(), index.properties() ) );
         }
+    }
+
+    public NodeCursor nodeCursor()
+    {
+        return nodeCursor;
+    }
+
+    public RelationshipScanCursor relationshipCursor()
+    {
+        return relationshipCursor;
+    }
+
+    public PropertyCursor propertyCursor()
+    {
+        return propertyCursor;
     }
 }
