+++ community/kernel/src/main/java/org/neo4j/kernel/impl/storageengine/impl/recordstorage/NodeValueClientFilter.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.impl.newapi;
+package org.neo4j.kernel.impl.storageengine.impl.recordstorage;
 
 import java.util.Arrays;
 import java.util.Comparator;
@@ -72,13 +72,13 @@ class NodeValueClientFilter implements NodeValueClient, IndexProgressor
     private final DefaultNodeCursor node;
     private final DefaultPropertyCursor property;
     private final IndexQuery[] filters;
-    private final Read read;
+    private final RecordStorageReader read;
     private int[] keys;
     private IndexProgressor progressor;
 
     NodeValueClientFilter(
             NodeValueClient target,
-            DefaultNodeCursor node, DefaultPropertyCursor property, Read read, IndexQuery... filters )
+            DefaultNodeCursor node, DefaultPropertyCursor property, RecordStorageReader read, IndexQuery... filters )
     {
         this.target = target;
         this.node = node;
