+++ community/kernel/src/main/java/org/neo4j/kernel/impl/storageengine/impl/recordstorage/References.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.impl.newapi;
+package org.neo4j.kernel.impl.storageengine.impl.recordstorage;
 
 import org.neo4j.kernel.impl.store.record.AbstractBaseRecord;
