+++ community/kernel/src/main/java/org/neo4j/kernel/impl/store/NodeStore.java
@@ -34,14 +34,13 @@ import org.neo4j.kernel.impl.store.record.NodeRecord;
 import org.neo4j.kernel.impl.store.record.RecordLoad;
 import org.neo4j.kernel.impl.util.Bits;
 import org.neo4j.logging.LogProvider;
-import org.neo4j.storageengine.api.StorageReader;
 
 import static org.neo4j.kernel.impl.store.NoStoreHeaderFormat.NO_STORE_HEADER_FORMAT;
 
 /**
  * Implementation of the node store.
  */
-public class NodeStore extends CommonAbstractStore<NodeRecord,NoStoreHeader> implements StorageReader.Nodes
+public class NodeStore extends CommonAbstractStore<NodeRecord,NoStoreHeader>
 {
     public static Long readOwnerFromDynamicLabelsRecord( DynamicRecord record )
     {
@@ -58,7 +57,6 @@ public class NodeStore extends CommonAbstractStore<NodeRecord,NoStoreHeader> imp
         return bits.getLong( requiredBits );
     }
 
-    @Override
     public RecordCursor<DynamicRecord> newLabelCursor()
     {
         return dynamicLabelStore.newRecordCursor( dynamicLabelStore.newRecord() ).acquire( getNumberOfReservedLowIds(),
