+++ community/kernel/src/main/java/org/neo4j/kernel/impl/store/RelationshipGroupStore.java
@@ -30,10 +30,8 @@ import org.neo4j.kernel.impl.store.id.IdGeneratorFactory;
 import org.neo4j.kernel.impl.store.id.IdType;
 import org.neo4j.kernel.impl.store.record.RelationshipGroupRecord;
 import org.neo4j.logging.LogProvider;
-import org.neo4j.storageengine.api.StorageReader;
 
-public class RelationshipGroupStore extends CommonAbstractStore<RelationshipGroupRecord,IntStoreHeader> implements
-        StorageReader.Groups
+public class RelationshipGroupStore extends CommonAbstractStore<RelationshipGroupRecord,IntStoreHeader>
 {
     public static final String TYPE_DESCRIPTOR = "RelationshipGroupStore";
