+++ community/kernel/src/test/java/org/neo4j/kernel/impl/api/KernelStatementTest.java
@@ -60,7 +60,7 @@ public class KernelStatementTest
         }
 
         // then
-        verify( storeStatement ).release();
+        verify( storeStatement ).endStatement();
     }
 
     @Test( expected = NotInTransactionException.class )
