+++ community/kernel/src/test/java/org/neo4j/kernel/impl/newapi/MockStore.java
@@ -50,6 +50,8 @@ import org.neo4j.io.pagecache.PageCursor;
 import org.neo4j.kernel.api.ExplicitIndex;
 import org.neo4j.kernel.api.exceptions.index.IndexNotFoundKernelException;
 import org.neo4j.kernel.impl.api.KernelTransactionImplementation;
+import org.neo4j.kernel.impl.storageengine.impl.recordstorage.DefaultCursors;
+import org.neo4j.kernel.impl.storageengine.impl.recordstorage.DefaultPropertyCursor;
 import org.neo4j.kernel.impl.store.DynamicRecordAllocator;
 import org.neo4j.kernel.impl.store.PropertyStore;
 import org.neo4j.kernel.impl.store.RecordCursor;
